final quicktest cleanups
Move the temporary qtmockusso and qtcandidtest packages
back to mockusso and candidtest respectively.

Also convert a couple of remaining packages to use quicktest
(missed because they didn't have package_test.go files).
