+++ cmd/candid/internal/admincmd/show_test.go
@@ -8,12 +8,12 @@ import (
 	"testing"
 	"time"
 
+	qt "github.com/frankban/quicktest"
+	"github.com/frankban/quicktest/qtsuite"
 	"golang.org/x/net/context"
 	"gopkg.in/macaroon-bakery.v2/bakery"
 
 	"github.com/CanonicalLtd/candid/store"
-	qt "github.com/frankban/quicktest"
-	"github.com/frankban/quicktest/qtsuite"
 )
 
 type showSuite struct {
