+++ internal/v1/users_test.go
@@ -21,9 +21,9 @@ import (
 	"github.com/CanonicalLtd/candid/idp"
 	testidp "github.com/CanonicalLtd/candid/idp/test"
 	"github.com/CanonicalLtd/candid/internal/auth"
+	"github.com/CanonicalLtd/candid/internal/candidtest"
 	"github.com/CanonicalLtd/candid/internal/discharger"
 	"github.com/CanonicalLtd/candid/internal/identity"
-	candidtest "github.com/CanonicalLtd/candid/internal/qtcandidtest"
 	"github.com/CanonicalLtd/candid/internal/v1"
 	"github.com/CanonicalLtd/candid/store"
 )
